feat! make SUC opt-in
SUC can cause issues when people are bootstrapping new clusters and there's a newer k3s version available. This can cause the cluster to fail or become unresponsive. This should be enabled with caution.

If you want to enable this, make sure you build your cluster with the latest k3s version set in the Ansible vars.